Opening and closing menu should now clear glitch material
i'm sure there are better ways to do this, but if there's any chance the glitch effect gets stuck on, this should allow the player to clear it by opening and closing the menu. this will cause brief strangeness during timelines where the glitch strength is being manipulated, but we can fix this once we have time, and in the meantime, maybe a cosmetic glitch is better than a potential game breaking one?